fixed #1

+++ src/NPSService.java
@@ -14,9 +14,9 @@ public class NPSService {
             }
             neutrals++;
         }
-        double all = detractors + promoters + neutrals;
-        if (all != 0.0){
-            return promoters*100/all - detractors*100/all;
+        int all = detractors + promoters + neutrals;
+        if (all != 0){
+            return (double)promoters*100/all - (double)detractors*100/all;
         }
         return 0.0;
     }
